Renamed 'label' package, fixed jpf-label related errors
