+++ code/jpf-ctl/src/main/java/labels/BinaryLabel.java
@@ -1,4 +1,4 @@
-package label;
+package labels;
 
 public abstract class BinaryLabel implements Label {
