+++ code/jpf-ctl/src/main/java/labels/End.java
@@ -1,4 +1,4 @@
-package label;
+package labels;
 
 public class End extends UnaryLabel{
