+++ code/jpf-ctl/src/main/java/labels/Generator.java
@@ -1,4 +1,4 @@
-package label;
+package labels;
 
 import java.util.List;
 import java.util.regex.Pattern;
