+++ code/jpf-ctl/src/main/java/labels/Initial.java
@@ -1,4 +1,4 @@
-package label;
+package labels;
 
 public class Initial extends UnaryLabel {
