+++ code/jpf-ctl/src/main/java/labels/IntegerStaticField.java
@@ -1,4 +1,4 @@
-package label;
+package labels;
 
 public class IntegerStaticField extends BinaryLabel {
