+++ code/jpf-ctl/src/main/java/labels/ReturnedIntegerMethod.java
@@ -1,4 +1,4 @@
-package label;
+package labels;
 
 public class ReturnedIntegerMethod extends BinaryLabel {
