+++ code/jpf-ctl/src/main/java/labels/ThrownException.java
@@ -1,4 +1,4 @@
-package label;
+package labels;
 
 public class ThrownException extends BinaryLabel {
