+++ code/jpf-ctl/src/main/java/labels/UnaryLabel.java
@@ -1,4 +1,4 @@
-package label;
+package labels;
 
 public abstract class UnaryLabel implements Label {
